esp32/machine_uart: Support different SoC's default UART pins.
Currently hardcoded values are used for the default pins for UART1, which
are only correct some ESP32 SoCs.  Others use different pins.

The pins on UART0 were not set to the defaults, but left as is.  Likely
this is because usually the REPL in on UART0 and it is assumed the REPL
UART will already be configured to the correct pins.

However, the REPL might not be on UART0.  Code later in this function
accounts for that, but not when the default pins are set.

The current code also assumes at most 3 UARTs.  The ESP32-P4 has 6.

The default gpio for each UART is available from ESP-IDF, so change the
code to use that.  This allows it to work for any number of UARTs and to
support all current/future SoCs.  And without ifdefs.

Don't set the pins to the default pin for MICROPY_HW_UART_REPL, if any,
rather than assuming UART0 is MICROPY_HW_UART_REPL.

Signed-off-by: Trent Piepho <tpiepho@gmail.com>