esp32/uart: Default UART REPL to same UART as ESP console
ESP-IDF defines a console UART, which is where log messages will go.
Usually the REPL is on this same UART.

The ESP-IDF will log messages at boot to the console at the baud rate it
is configured for, and then micropython will take over and switch to its
baud rate.  It doesn't make much sense for these baud rates to be
different.  There's no way for the recevier to know when to switch.

So, instead of assuming 115200 baud, make the default micropython UART
REPL baud rate match the ESP-IDF console baud rate (which is a build
time constant).  A micropython board configuration can still specify a
different rate if it wants to.