esp32/modmachine: Support for DFS and auto light sleep in machine.freq
Extend machine.freq() to have the signature:

machine.freq([hz, min_hz=hz, lightsleep=False])

If a min_hz value is specified, it will be use to enable DFS.  The default
when no min_hz is specified is to set it to the same as the hz value, which
was the previous behavior and effectively disables DFS.

If lightsleep is to set True, then automatic light sleep will be enabled.

These features require support in the ESP-IDF configuration used.