esp32/uart: Wake from lightsleep on stdin.
The current code busy-polls for input from stdin.  This means it can't
sleep when automatic lightsleep is enabled.

Have the the main task wait on an empty stdin ringbuf indefintely rather
than 1 tick only.  Have the stdin uart ISR notify the main task so it
wakes on input.

Then there is the problem that the UART doesn't work in lightsleep!  First,
enable wake on UART.  This will produce a UART wake interrupt when enough
rising edges on the RX line are detected.  It doesn't receive the character
that was sent, just wakes from sleep.  The data is lost.

Then add a period after being woken or receiving uart input when lightsleep
is disabled, via a PM lock.  If this wasn't done, auto lihtsleep would
immediatly go back to sleep after being woken on UART and never sucessfully
receive a character.

The timeouts are hardcoded in defines, but it could be set at runtime.
There are two, one after receiving data and one after waking but not yet
receiving any data.  The latter can be shorter since there are reasons to
wake other than UART input and there isn't an event mechanism that can tell
the difference.