esp32/machine_pin: Allow RTC mode pull control.
In lightsleep or deepsleep, the normal pull up/down control is turned
off.  But the RTC power domain can still supply pull.  However, this
must be enabled seperately from the normal power domain pull control (on
most ESP32 MCUs).

This adds an rtcpull boolean option to enable pull in the RTC power
domain that matches the normal pull.  On some ESP32 MCUs, the RTC pull
and the normal pull can be different, but this isn't supported.